Warning fixes, style

+++ src/BattleSystem.cpp
@@ -69,7 +69,7 @@ void BattleSystem::handleAttacks()
 void BattleSystem::debugDraw()
 {
     auto &rnd = m_app.getRenderer();
-    for (int i = 0; i < m_appliedHits.getFilled(); ++i)
+    for (size_t i = 0; i < m_appliedHits.getFilled(); ++i)
     {
         const auto &el = m_appliedHits[i];
         rnd.drawCross(el, {1, 5}, {5, 1}, {0, 0, 0, 255}, m_cam);
